requirements.txt
Package                   Version
------------------------- ---------
altair                    5.2.0
attrs                     23.2.0
blinker                   1.7.0
cachetools                5.3.3
certifi                   2024.2.2
charset-normalizer        3.3.2
click                     8.1.7
colorama                  0.4.6
contourpy                 1.2.0
cycler                    0.12.1
et-xmlfile                1.1.0
fonttools                 4.48.1
gitdb                     4.0.11
GitPython                 3.1.42
google-api-core           2.18.0
google-auth               2.29.0
google-cloud              0.34.0
google-cloud-core         2.4.1
google-cloud-storage      2.16.0
google-crc32c             1.5.0
google-resumable-media    2.7.0
googleapis-common-protos  1.63.0
idna                      3.6
Jinja2                    3.1.3
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
kiwisolver                1.4.5
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.8.3
mdurl                     0.1.2
mplfinance                0.12.10b0
numpy                     1.26.4
openpyxl                  3.1.2
packaging                 23.2
pandas                    2.2.0
pillow                    10.2.0
pip                       23.2.1
plotly                    5.20.0
proto-plus                1.23.0
protobuf                  4.25.3
pyarrow                   15.0.2
pyasn1                    0.6.0
pyasn1_modules            0.4.0
pydeck                    0.8.1b0
Pygments                  2.17.2
pyparsing                 3.1.1
python-dateutil           2.8.2
pytz                      2024.1
referencing               0.34.0
requests                  2.31.0
rich                      13.7.1
rpds-py                   0.18.0
rsa                       4.9
six                       1.16.0
smmap                     5.0.1
streamlit                 1.32.2
tenacity                  8.2.3
toml                      0.10.2
toolz                     0.12.1
tornado                   6.4
typing_extensions         4.10.0
tzdata                    2024.1
urllib3                   2.2.1
watchdog                  4.0.0
